BackEnd/requirements.txt
fastapi
uvicorn
firebase-admin
sqlalchemy
pyodbc
pydantic
pydantic-settings
httpx
python-dotenv
pydantic[email]